[VBV-1960] Disallow global registries to be deleted from project registries view
Currently, it is possible to select any registry in the project
registries view (regardless of its scope) and issue a delete command for
it. If you are the cloud admin, this request will succeed, if not, it
will fail with 403 and the red invalid session header will pop out.
Global registries are meant to be managed by cloud admins only and
through the Global registries view only. They are displayed in the
project registries view for info purposes only.

With this commit, it is now not possible to issue delete requests for global
registries from the project registries view. If the delete button is
pressed with at least one global registry in the current selection, a
warning message will be displayed to clarify that deletion of these
registries was skipped. Also, the delete confirmation popup now notes
that global registries cannot be deleted from this view.

VIDEO PREVIEW: http://go/5j7f

Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/31468
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>

+++ ui/ng-app/src/app/views/projects/project-details/project-registries.component.ts
@@ -15,6 +15,7 @@ import { DocumentListResult, DocumentService } from "../../../utils/document.ser
 import { ErrorService } from "../../../utils/error.service";
 import { CancelablePromise, Utils} from "../../../utils/utils";
 import { Links } from "../../../utils/links";
+import * as I18n from 'i18next';
 
 @Component({
     selector: 'app-project-registries',
@@ -25,6 +26,10 @@ import { Links } from "../../../utils/links";
  * View for Project Registries.
  */
 export class ProjectRegistriesComponent implements OnInit {
+    tabAlertClosed: boolean = true;
+    tabAlertType: string = 'alert-warning';
+    tabAlertText: string;
+
     loading: boolean;
     loadingPromise: CancelablePromise<DocumentListResult>;
 
@@ -52,6 +57,12 @@ export class ProjectRegistriesComponent implements OnInit {
 
     ngOnDestroy() {
         this.sub.unsubscribe();
+    }
+
+    private warnGlobalRegistriesSkippedDeletion(): void {
+        this.tabAlertType = 'alert-warning';
+        this.tabAlertText = I18n.t('projects.projectRegistries.globalRegistriesDeletionSkipped.warningMessage');
+        this.tabAlertClosed = false;
     }
 
     listProjectRegistries(queryOptions) {
@@ -93,16 +104,24 @@ export class ProjectRegistriesComponent implements OnInit {
 
     deleteConfirmed() {
         let promises: any[] = [];
+        let globalRegistriesSelected: boolean = false;
 
         this.selectedProjectRegistries.forEach((registriesToDelete) => {
-           let deletePromise =  this.service.delete(registriesToDelete.documentSelfLink);
-           promises.push(deletePromise);
+            if (this.isProjectSpecificRegistry(registriesToDelete)) {
+                let deletePromise =  this.service.delete(registriesToDelete.documentSelfLink);
+                promises.push(deletePromise);
+            } else {
+                globalRegistriesSelected = true;
+            }
         });
 
         Promise.all(promises).then(() => {
             this.selectedProjectRegistries = [];
             this.showDeleteConfirmation = false;
             this.listProjectRegistries({});
+            if (globalRegistriesSelected) {
+                this.warnGlobalRegistriesSkippedDeletion();
+            }
         }).catch(err => {
             this.deleteConfirmationAlert = Utils.getErrorMessage(err)._generic;
         });
@@ -114,5 +133,13 @@ export class ProjectRegistriesComponent implements OnInit {
 
     getRegistryId(registry){
         return Utils.getDocumentId(registry.documentSelfLink);
+    }
+
+    isProjectSpecificRegistry(registry): boolean {
+        return registry
+            && registry.tenantLinks
+            && this.projectLink
+            && Array.isArray(registry.tenantLinks)
+            && registry.tenantLinks.includes(this.projectLink);
     }
 }
